HW REV: Fix stability issue
Important change required to ensure stability across different types of ethernet switches.

I had only tested this with a TP-LINK switch, but the design did not work with other switches, and the reference schematic does NOT specify that the CT/termination resistors should be tied to 3.3V. Instead they should be coupled to GND through a capacitor (of unspecified value).